wip: minimal execution engine stub
