+++ mod/state-transition/pkg/core/state_processor_genesis_test.go
@@ -21,19 +21,11 @@
 package core_test
 
 import (
-	"context"
-	"fmt"
 	"testing"
 
-	corestore "cosmossdk.io/core/store"
-	"cosmossdk.io/log"
-	"cosmossdk.io/store"
-	"cosmossdk.io/store/metrics"
-	storetypes "cosmossdk.io/store/types"
 	"github.com/berachain/beacon-kit/mod/config/pkg/spec"
 	"github.com/berachain/beacon-kit/mod/consensus-types/pkg/types"
 	engineprimitives "github.com/berachain/beacon-kit/mod/engine-primitives/pkg/engine-primitives"
-	"github.com/berachain/beacon-kit/mod/node-core/pkg/components"
 	"github.com/berachain/beacon-kit/mod/primitives/pkg/common"
 	"github.com/berachain/beacon-kit/mod/primitives/pkg/crypto/mocks"
 	"github.com/berachain/beacon-kit/mod/primitives/pkg/math"
@@ -42,10 +34,6 @@ import (
 	"github.com/berachain/beacon-kit/mod/state-transition/pkg/core"
 	statedb "github.com/berachain/beacon-kit/mod/state-transition/pkg/core/state"
 	"github.com/berachain/beacon-kit/mod/storage/pkg/beacondb"
-	"github.com/berachain/beacon-kit/mod/storage/pkg/db"
-	"github.com/berachain/beacon-kit/mod/storage/pkg/encoding"
-	dbm "github.com/cosmos/cosmos-db"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 )
@@ -89,7 +77,7 @@ type (
 
 func TestInitialize(t *testing.T) {
 	cs := spec.TestnetChainSpec()
-	// in.ExecutionEngine,
+	execEngine := &testExecutionEngine{}
 	mocksSigner := &mocks.BLSSigner{}
 
 	sp := core.NewStateProcessor[
@@ -112,7 +100,7 @@ func TestInitialize(t *testing.T) {
 		types.WithdrawalCredentials,
 	](
 		cs,
-		nil,
+		execEngine,
 		mocksSigner,
 	)
 
@@ -152,66 +140,4 @@ func TestInitialize(t *testing.T) {
 	)
 	require.NoError(t, err)
 	require.Len(t, vals, len(deposits))
-}
-
-// Unit tests helpers
-
-type testKVStoreService struct {
-	ctx sdk.Context
-}
-
-func (kvs *testKVStoreService) OpenKVStore(context.Context) corestore.KVStore {
-	//nolint:contextcheck // fine with tests
-	return components.NewKVStore(
-		sdk.UnwrapSDKContext(kvs.ctx).KVStore(testStoreKey),
-	)
-}
-
-var (
-	testStoreKey = storetypes.NewKVStoreKey("state-transition-tests")
-	testCodec    = &encoding.SSZInterfaceCodec[*types.ExecutionPayloadHeader]{}
-)
-
-func initTestStore() (
-	*beacondb.KVStore[
-		*types.BeaconBlockHeader,
-		*types.Eth1Data,
-		*types.ExecutionPayloadHeader,
-		*types.Fork,
-		*types.Validator,
-		types.Validators,
-	], error) {
-	db, err := db.OpenDB("", dbm.MemDBBackend)
-	if err != nil {
-		return nil, fmt.Errorf("failed opening mem db: %w", err)
-	}
-	var (
-		nopLog     = log.NewNopLogger()
-		nopMetrics = metrics.NewNoOpMetrics()
-	)
-
-	cms := store.NewCommitMultiStore(
-		db,
-		nopLog,
-		nopMetrics,
-	)
-
-	ctx := sdk.NewContext(cms, true, nopLog)
-	cms.MountStoreWithDB(testStoreKey, storetypes.StoreTypeIAVL, nil)
-	if err = cms.LoadLatestVersion(); err != nil {
-		return nil, fmt.Errorf("failed to load latest version: %w", err)
-	}
-	testStoreService := &testKVStoreService{ctx: ctx}
-
-	return beacondb.New[
-		*types.BeaconBlockHeader,
-		*types.Eth1Data,
-		*types.ExecutionPayloadHeader,
-		*types.Fork,
-		*types.Validator,
-		types.Validators,
-	](
-		testStoreService,
-		testCodec,
-	), nil
 }
